Abstract away the parser
