+++ browser.js
@@ -1,41 +1,3 @@
-var browsers = [
-  [ 'edge', /Edge\/([0-9\._]+)/ ],
-  [ 'chrome', /Chrom(?:e|ium)\/([0-9\.]+)(:?\s|$)/ ],
-  [ 'firefox', /Firefox\/([0-9\.]+)(?:\s|$)/ ],
-  [ 'opera', /Opera\/([0-9\.]+)(?:\s|$)/ ],
-  [ 'ie', /Trident\/7\.0.*rv\:([0-9\.]+)\).*Gecko$/ ],
-  [ 'ie', /MSIE\s([0-9\.]+);.*Trident\/[4-7].0/ ],
-  [ 'ie', /MSIE\s(7\.0)/ ],
-  [ 'bb10', /BB10;\sTouch.*Version\/([0-9\.]+)/ ],
-  [ 'android', /Android\s([0-9\.]+)/ ],
-  [ 'ios', /iPad\;\sCPU\sOS\s([0-9\._]+)/ ],
-  [ 'ios',  /iPhone\;\sCPU\siPhone\sOS\s([0-9\._]+)/ ],
-  [ 'safari', /Safari\/([0-9\._]+)/ ]
-];
+const detectBrowser = require('./lib/detectBrowser');
 
-var i = 0, mapped =[];
-for (i = 0; i < browsers.length; i++) {
-  browsers[i] = createMatch(browsers[i]);
-  if (isMatch(browsers[i])) {
-    mapped.push(browsers[i]);
-  }
-}
-
-var match = mapped[0];
-var parts = match && match[3].split(/[._]/).slice(0,3);
-
-while (parts && parts.length < 3) {
-  parts.push('0');
-}
-
-// set the name and version
-exports.name = match && match[0];
-exports.version = parts && parts.join('.');
-
-function createMatch(pair) {
-  return pair.concat(pair[1].exec(navigator.userAgent));
-}
-
-function isMatch(pair) {
-  return !!pair[2];
-}
+module.exports = detectBrowser(navigator.userAgent);
